Added untested pitch frequency detector based on Cepstrum method with DWT peak selection.
Relies on Aquila. Download Aquila, install it with cmake etc. Make sure
ooura.dll is located in Aquila/lib. If not, compile fft4g.c in the
lib/ooura subfolder in the source directory as a shared library and copy
it over. Update library and include paths as necessary.

TODO: Someone please make the configuration and installation easier.